# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Rush Hour Real Estate Platform API
  version: 1.0.0
  description: MVP платформы недвижимости - публичные эндпоинты
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
  /areas:
    get:
      operationId: listAreas
      summary: Получить список районов
      tags:
        - Areas
      parameters:
        - name: include
          in: query
          schema:
            type: string
            enum: [boundary]
          description: Включить границы районов (GeoJSON)
      responses:
        "200":
          description: Список районов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Area"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /projects:
    get:
      operationId: listProjects
      summary: Получить список проектов
      tags:
        - Projects
      parameters:
        - name: area
          in: query
          schema:
            type: string
          description: Фильтр по slug района
      responses:
        "200":
          description: Список проектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /lots:
    get:
      operationId: listLots
      summary: Получить список лотов (каталог)
      tags:
        - Lots
      parameters:
        - name: area
          in: query
          schema:
            type: string
          description: Фильтр по slug района
        - name: project
          in: query
          schema:
            type: string
          description: Фильтр по slug проекта
        - name: type
          in: query
          schema:
            type: string
            enum: [apartment, villa, townhouse, penthouse]
          description: Тип лота
        - name: bedrooms
          in: query
          schema:
            type: integer
          description: Количество спален
        - name: priceMin
          in: query
          schema:
            type: number
          description: Минимальная цена
        - name: priceMax
          in: query
          schema:
            type: number
          description: Максимальная цена
        - name: areaMin
          in: query
          schema:
            type: number
          description: Минимальная площадь (м²)
        - name: areaMax
          in: query
          schema:
            type: number
          description: Максимальная площадь (м²)
        - name: bonus
          in: query
          schema:
            type: string
          description: Фильтр по ключу бонуса (например, goldenVisa)
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, area_desc, newest]
            default: newest
          description: Сортировка
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Номер страницы
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Количество элементов на странице
      responses:
        "200":
          description: Список лотов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LotsListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
    Area:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        city:
          type: string
        lat:
          type: number
        lng:
          type: number
        status:
          type: string
          enum: [active, inactive]
        data:
          $ref: "#/components/schemas/AreaData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AreaData:
      type: object
      properties:
        boundary:
          $ref: "#/components/schemas/GeoJSONPolygon"
        center:
          $ref: "#/components/schemas/Point"
        zoom:
          type: integer
        bbox:
          $ref: "#/components/schemas/BoundingBox"
        seo:
          type: object
          additionalProperties: true
    GeoJSONPolygon:
      type: object
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
                minItems: 2
                maxItems: 2
    Point:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
    BoundingBox:
      type: object
      properties:
        southWest:
          $ref: "#/components/schemas/Point"
        northEast:
          $ref: "#/components/schemas/Point"
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, archived]
        developerId:
          type: string
          format: uuid
        areaId:
          type: string
          format: uuid
        lat:
          type: number
        lng:
          type: number
        data:
          $ref: "#/components/schemas/ProjectData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lots:
          type: array
          items:
            $ref: "#/components/schemas/Lot"
          description: Первые N лотов проекта (если запрошено через includeLots)
    ProjectData:
      type: object
      properties:
        description:
          oneOf:
            - type: string
            - type: object
              additionalProperties:
                type: string
        specs:
          type: object
          additionalProperties: true
        featuresAmenities:
          type: array
          items: {}
        media:
          $ref: "#/components/schemas/Media"
    Media:
      type: object
      properties:
        cover:
          $ref: "#/components/schemas/MediaItem"
        gallery:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
    MediaItem:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
    Lot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, hidden, reserved, sold]
        projectId:
          type: string
          format: uuid
        developerId:
          type: string
          format: uuid
        areaId:
          type: string
          format: uuid
        type:
          type: string
          enum: [apartment, villa, townhouse, penthouse]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        areaSqm:
          type: number
        floor:
          type: integer
        priceCurrency:
          type: string
          default: AED
        priceAmount:
          type: number
        bonusKeys:
          type: array
          items:
            type: string
        data:
          $ref: "#/components/schemas/LotData"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LotData:
      type: object
      properties:
        media:
          $ref: "#/components/schemas/LotMedia"
        paymentPlan:
          $ref: "#/components/schemas/PaymentPlan"
        bonuses:
          type: array
          items:
            $ref: "#/components/schemas/Bonus"
        floorPosition:
          $ref: "#/components/schemas/FloorPosition"
        tags:
          type: array
          items:
            type: string
    LotMedia:
      type: object
      properties:
        photos:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        floorPlanImages:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        cover:
          $ref: "#/components/schemas/MediaItem"
    PaymentPlan:
      type: object
      properties:
        schedule:
          type: array
          items:
            $ref: "#/components/schemas/PaymentScheduleItem"
    PaymentScheduleItem:
      type: object
      properties:
        stage:
          type: string
        percent:
          type: number
        amount:
          type: number
        dueDate:
          type: string
    Bonus:
      type: object
      properties:
        title:
          type: string
        style:
          type: string
        description:
          type: string
    FloorPosition:
      type: object
      properties:
        label:
          type: string
        x:
          type: number
        y:
          type: number
    LotsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Lot"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              oneOf:
                - type: string
                - type: object
                  additionalProperties: true
  responses:
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "internal_error"
              message: "Internal server error"
...
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: API ключ для доступа к админ-эндпоинтам